Moving braket module inside objective
